Add timeseries normalization.
The normalization permits to relocate each point at a given interval
from each others (60 seconds in our case). It does not interpolate the
data, instead it is based on the amount of data. For example, if you
have a time serie representing the amount of data transfered at a given
time, the normalization will recalculate the amount of data transfered
at the wanted interval by computing the average of all points in the
interval.

The normalization has been implemented in a non destructive way by
simply
creating a new type of iterator which returns each point aligned
according
to an interval. It can be activated through the web ui or the cli.

More information and example can be found here:
http://www.vandenbogaerdt.nl/rrdtool/process.php